+++ src/examples/example.module.ts
@@ -144,6 +144,7 @@ import { ExampleDataRowAddComponent } from './data-row-add/example-data-row-add.
 import { ExampleColumnTooltipBasicComponent } from './column-tooltip-basic/example-column-tooltip-basic.component';
 import { ExampleDataRowDeleteComponent } from './data-row-delete/example-data-row-delete.component';
 import { ExampleDetailsRowPinComponent } from './details-row-pin/example-details-row-pin.component';
+import { ExampleSelectRowDisableComponent } from './select-row-disable/example-select-row-disable.component';
 
 const EXAMPLES: Array<any> = [
 	ExampleActionBarBasicComponent,
@@ -266,6 +267,7 @@ const EXAMPLES: Array<any> = [
 	ExampleSelectColumnBasicComponent,
 	ExampleSelectMixBasicComponent,
 	ExampleSelectRowBasicComponent,
+	ExampleSelectRowDisableComponent,
 	ExampleSelectRowCommandComponent,
 	ExampleSelectRowSingleComponent,
 	ExampleSizeColumnAbsoluteComponent,