+++ src/lib/pipes/pipe.module.ts
@@ -4,11 +4,13 @@ import { FilterPipe } from './filter.pipe';
 import { TimePipe } from './time.pipe';
 import { NumberPipe } from './number.pipe';
 import { DatePipe } from './date.pipe';
+import { TextPipe } from './text.pipe';
 import { CurrencyPipe } from './currency.pipe';
 import { ArrayPipe } from './array.pipe';
 
 @NgModule({
 	declarations: [
+		TextPipe,
 		HighlightPipe,
 		FilterPipe,
 		TimePipe,
@@ -18,6 +20,7 @@ import { ArrayPipe } from './array.pipe';
 		ArrayPipe,
 	],
 	exports: [
+		TextPipe,
 		HighlightPipe,
 		FilterPipe,
 		TimePipe,